# to-do list crud api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "version": "1.0",
    "description": "A small SQLite-backed CRUD API for managing tasks."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Get API info",
        "description": "Returns the API name, version, and available endpoints.",
        "responses": {
          "200": {
            "description": "API metadata"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Get health status",
        "description": "Returns a simple ok response to show the server is alive.",
        "responses": {
          "200": {
            "description": "Server is healthy"
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List tasks",
        "description": "Returns all tasks currently stored in the database.",
        "responses": {
          "200": {
            "description": "List of tasks"
          },
          "400": {
            "description": "Invalid query parameter"
          }
        }
      },
      "post": {
        "summary": "Create a task",
        "description": "Creates a new task from a title sent in the JSON body.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Buy milk"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created"
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a task",
        "description": "Returns one task by id or a 404 error if it does not exist.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },
      "put": {
        "summary": "Update a task",
        "description": "Updates the title and/or done value for an existing task.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "done": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated"
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "description": "Deletes a task by id and returns no content.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    }
  }
}